MB-39684 Add new PAUSED state to runtie operator
When an intersect scan is under the inner of a nested-loop join,
when one of the scans finishes it sends a stop to the other scans.
In case another scan is still in _CREATED state when this happens,
it'll be set to _KILLED state. Then when the next outer comes and
a reopen() is called on the inner, the reopen() call returns false
and the inner scan is terminated, resulting in early termination
of the nested-loop join, leading to wrong result.

This checkin introduces a new _PAUSED state, which is very much
like the _KILLED state except that it can be safely reopened
(since it was in _CREATED state before). In intersect scan, when
one of the scans finishes, it sends a PAUSE call (SendPause())
instead of a STOP call (SendStop()), such that the other scans
can be set to _PAUSED if it was in _CREATED state. All other
functionality remains the same between _PAUSED and _KILLED
except for the ability to be safely reopened. This new
mechanism is only performed when the intersect scan is under
inner of a nested-loop join.

Change-Id: Id143367731a4c15b0117f70d339e6ed36c510b42
Reviewed-on: http://review.couchbase.org/c/query/+/130495
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>