MB-39988 Put filter evaluation after Fetch operator
Part 1 of the changes for early filtering put a filter inside
the Fetch operator if it's not a covering index scan. The Fetch
operator is alreay heavily loaded with potential unmarshal of
the documents and thus is slow. By putting additional filter
evaluation inside the Fetch operator, especially if the filter
is non-simple (involving functions, etc), we are in effect
serializing more, and may not be able to take advantage of
multiple cpus when available.

In case there is a filter, instead of putting it inside the
Fetch operator, use a separate Filter operator after the
Fetch. This way the filter evaluation can potentially be
done in parallel, and even when parallelism is not enabled,
since it runs in a separate go thread, we can still run
Fetch and Filter on separate cpus at the same time.

This change is only done for the Fetch operator. Since adding
a separate Filter operator also has it associated costs. For
other operators that also has a filter (such as covering
index scan, joins, unnest, etc) leave those as is for now.

Change-Id: I2a14ea8dc6c9fa556af0a01b097029f24f4cd415
Reviewed-on: http://review.couchbase.org/c/query/+/130932
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>