MB-40072 Get document count from dictionary cache for intersect scan
Currently in costing of intersect scan, we have a Count() call on the
underlying keyspace to get document count, this call "fails"
occasionally and produces 0 as document count, this then triggers
the optimizer to calculate negative costs and then fall back to
nested-loop join.

We already have document count as part of dictionary so instead of
calling Count() directly on the keyspace we simply get the same
information from dictionary.

A separate bug will be filed to address the actual issue of Count()
call returning 0 with no error returned.

Change-Id: I371493474ebdff931243a8262354c2cfa5f8ac88
Reviewed-on: http://review.couchbase.org/c/query/+/131116
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>