MB-37115 Allow two expressions to be considered equivalent when binding variable names differ
For an ANY clause, or an ARRAY construct, current code only considers two such
expressions to be equivalent when everything is the same. In practice there are
often situations where the binding variable names used are different but
everything else is the same. Previously a rename operation is required before
comparing such expressions.

Enhance the logic for EquivalentTo() for collPred and collMap in expression
package such that there is an option for strict comparison (everything needs
to be the same) or non-strict comparison (allow binding variable names
to be different). With non-strict comparison a rename operation is done
inside the comparison function such that the caller can be much simpler.

Also renamed the interface type expression.CollectionPredicate to
expression.CollPredicate, to avoid confusion with collections under bucket
and scope. The n1fty repo currently uses expression.CollectionPredicate,
so a stub is left in place, until n1fty renamed its reference as well.

Change-Id: I32fb48e66439e95dbdd75f775b8d42427ad0eeec
Reviewed-on: http://review.couchbase.org/c/query/+/131690
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>