MB-40759 Initialize dictionary cache before accepting requests
The fix for MB-39839 moved initialization of dictionary cache to
after a server is created, which means in an overloaded system
we could be potentially accepting requests before the dictionary
cache is initialized. Since the original issue from MB-39839 is
being handled differently we move the dictionary cache
initialization back to its original location.

Change-Id: I48db6e905fa5f9e81052829bd8941eb4403e6d58
Reviewed-on: http://review.couchbase.org/c/query/+/133851
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>