MB-40931 load indicator in metrics for SDK
in order to aid the SDK to choose which node to use, the metrics will also have an indicator of the load in the query node that has executed the request.

sample output:

...
"status": "success",
"metrics": {"elapsedTime": "128.30799ms","executionTime": "128.137889ms","resultCount": 4096,"resultSize": 4221869,"serviceLoad": 2}
}

In the first implementation this is going to be a number from 0 (no load) passing to 100 (all servicers busy), so several hundreds (all servicers busy
plus requests in queue).
SDK team will run with it on an environment simulating a customer scenario, and then we will amend
in whichever way it returns the most usable results.

Change-Id: I462d59d707e9a54ab0278ecba798d30265231c7c
Reviewed-on: http://review.couchbase.org/c/query/+/134469
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>