MB-41343 fix prepared prime and other clustered operations when node not yet added to cluster
CBSE complains that cust is removing nodes and re-adding them with delayed recovery
At start up N1QL can't determine that it's in a cluster and does not prime the prepared cache
Similar issues with functions storage access, until the node is rebalanced in.

Record multiple states for the node, delay node priming until node added to the cluster,
and also try to determine that the node is in fact started by ns_server and guess
node name, so that functions and prepared work without hiccups / performance issues
from the beginning, and not just after the node is balanced in.

Change-Id: Iac1b074882f3317eaf2ebfb54fe1f4f924784833
Reviewed-on: http://review.couchbase.org/c/query/+/135788
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>