MB-37686 MB-38720 race conditions in SendStop and Done
The proper change would probably be to change all of the SendAction()s not to send anything if baseSendAction() determines that
the opState is _COMPLETED, _KILLED, _PANICKED, _STOPPED - however that would open a different can of worms, in the sense
that we could possibly not send actions that should be still sent, because the children are still alive.

Hopefully, this defensive change might fix it.
If not, I'll have to have a look at more robust changes.

Change-Id: I743cee0649d6b6a8acc4985370d56e4a7a72814a
Reviewed-on: http://review.couchbase.org/c/query/+/136827
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>