MB-41486 annotated copy performance dip
In order to counter the unnest performance hit, we now have to annotated value methods to set anotations:
CopyAnnotations(), when it's unsafe to share them (eg update), and ShareAnnotations(), for when
it's safe to do so (eg setting an alias in a parent value).

For shared annotation, this is marked explicitly, so that they are handled properly when recycling values.

In time, as we are improving on expression, we could move collection expression annotation copies to shares.

Change-Id: I0592e0e9dba38968d27aa9b686d761f0b33ce939
Reviewed-on: http://review.couchbase.org/c/query/+/137261
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>