MB-41306 pool meta attachment
This change pools metas as well as attachments, and separates metas and attachments, so that SetAttachment("meta") and GetAttachment("meta")
are no longer needed.
This avoids allocating attachments if only meta is needed.

It also optimizes scopes to create and pool an object that can then be drill to pool nested scope and annotated values.
It fixes an issue where nested values weren't tracked, and could have given rise to panics if recycled.
This is currently only used for document keys, but it could be later extended.

With this change, Q26 pools 2 more maps (1 object value and one meta), and one more annotated value.
Q26 99th percentile latency goes down from 4.8s to 3.8s and memory usage from 25G to 16G.
There are marginal improvements in Q1 and Q2.

Change-Id: I366d2bbbb37a7f9d2ae920da06584b74b0fdc4fb
Reviewed-on: http://review.couchbase.org/c/query/+/137398
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>