MB-42187 Intersect scan used on two indexes with same keys in different order
In function narrowerOrEquivalent(), in mad-hatter we have the following logic
at the end of the function:

	if te.minKeys == nfcmatch || (shortest && te.minKeys == (nfcmatch+nkmatch)) {
		return true
	}

	return se.sumKeys > te.sumKeys ||
		(shortest && ((len(se.keys) <= len(te.keys)) ||
			(se.cond != nil && te.cond == nil && se.minKeys == te.minKeys)))

In cheshire-cat significant changes have been made to this function, however
we are missing the logic that checks for te.minKeys == (nfcmatch+nkmatch)
and returns true in that case. This check is in front of the checks on sumKeys,
len(keys) etc.

The fix is to add a check for nSargKeys == (snk + snc) and return true
(true in this case means equivalent). This check is added in front of the
checks for sumKeys, len(keys) etc to be consistent with previous version.
Also we now checks for nSargKeys instead of minKeys due to potential
skip keys. Note this fix also makes the previous fix for MB-42039 redundant.

Change-Id: I65ccdb06587f9a5f8c8ace6d5419a46cadff549f
Reviewed-on: http://review.couchbase.org/c/query/+/138609
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>