MB-41451 serialize sendAction() and Done
An OpStop() can come from a close or a timeout, which would be an asynchronous even compared to Done()
Rather than trying to make OpStop() and Done() work together, which is difficult since OpStop() uses structures that Done() is tearing down,
serialize OpStop() and Done().
Also, make sure that once an operator is killed, it can't try and stop anyone else.
As an added precaution, prevent SendAction() from sending actions if the operator has been torn down, and stop SendAction() midway if the children slice has been nillified

Change-Id: I7990efd27411d2b8e1c49411c9684f4eac34dfe4
Reviewed-on: http://review.couchbase.org/c/query/+/138919
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>