MB-42391 Proper handling of newly created SYS_CBO_STATS in dictionary
On cbq-engine startup, the query node checks for existence of the
system collection SYS_CBO_STATS, and set a flag on the local
dictionary to avoid repeated checks. The issue with this bug is
that if SYS_CBO_STATS does not exist initially, but gets automatically
created by UPDATE STATISTICS command, then we fail to update
the local flag to indicate SYS_CBO_STATS is now available, resulting
in symptoms such as a later delete not working. In addition, if the
cluster has multiple query nodes, and the first UPDATE STATISTICS
command was run on a remote query node, we also fail to update the
same flag on a different query node, resulting in similar issues.

The fix has two aspects:

1. When SYS_CBO_STATS is created (from UPDATE STATISTICS), also
   set the local flag to indicate that SYS_CBO_STATS is available

2. When a node performs UPDATE STATISTICS, it sends a DELETE
   request to remote query nodes to remote the corresponding entry
   in the dictionaries of remote nodes, such that on nect access
   it can be reloaded from SYS_CBO_STATS and thus get the new
   histograms just created. Enhance the logic to also detect
   cases when the dictionary cache is empty and re-perform the
   check for existence of SYS_CBO_STATS such that dictionary
   info on remote nodes are also up-to-date.

Change-Id: I97719044d1d5f129c0ae9d4a2369af4718486529
Reviewed-on: http://review.couchbase.org/c/query/+/139317
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>