MB-42467 lexer Stop() doesn't stop when error much behind shifted token
The issue is that the Stop() mechanism works if there's only one error in the chain,
because the stop will be picked up when the next token is read, and the first error happens
after a lot of tokens have been picked.

goyacc doesn't offer a YYABORT, but returning straight away in the rule seems to do the trick (although it's more of a hack than the previous hack).
Also, give full path of missing functions, so as to avoid all doubts.

Change-Id: Ifc277495b241d9c26c2355318e299814b67dc925
Reviewed-on: http://review.couchbase.org/c/query/+/139617
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>