MB-42474 Set exact spans properly in case of multiple spans for the same index
The problematic query has a predicate with OR clause and uses the same index
with multiple spans. On one side of the OR we use 2 sargable keys and on the
other side of the OR we use 1 sargable key. Both sides have exact span set
to true, however there is a different predicate on the OR arm which is not
considered (since on that side only a single sargable key is available),
but we did not detect it since when we determine on index pushdown later on
we take the max of the sargable keys from all arms of the OR clause.

To fix this, we add logic in sargForOR, when we iterate through different arms
of the OR clause, borrow logic from index pushdown determination to properly
detect whether there is a predicate that is not sargable, in which case
set exact span to be false.

Change-Id: I147a0d94e9bcc21d8d2d5643db882a866f24b304
Reviewed-on: http://review.couchbase.org/c/query/+/139700
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>