MB-42620 Panic with Advisor function with CBO-based query advise
When CBO is enabled (by default) and UPDATE STATISTICS has been
run on a keyspace (e.g. `travel-sample`), if an Advisor session
is started and stopped, internally the Advisor() funtion issues
Advise statements on the queries collected during the Advisor
session. The Advise statement uses CBO-based index advise since
UPDATE STATISTICS has been run on `travel-sample`, and it issues
internal UPDATE STATISTICS command as part of the Advise.

Advise statement requires a context to be set up. This is done
for an user-issued Advise statment but not done for an internal
Advise statement (from the Advisor function). The fix is to
also setup proper context for internally issued Advise statement
as well.

Change-Id: Ibcdcbd19f57bebdc832fb2d9a006d4be3ba954b0
Reviewed-on: http://review.couchbase.org/c/query/+/139853
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>