MB-42578 Allow subquery expression for INDEX option of UPDATE STATISTICS
Similar to BUILD INDEX statement where a subquery can be used to
get all indexes for a keyspace in "deferred" state, this checkin
allows a subquery to be used to get all GSI indexes in "online"
state for a keyspace, as argument to the INDEX option for
UPDATE STATISTICS, such that user does not need to list the names
of the indexes explicitly.

UPDATE STATISTICS FOR `travel-sample` INDEX ((SELECT RAW name FROM system:indexes WHERE keyspace_id = "travel-sample" AND state = "online" AND `using` = "gsi"));

or for a collection:

UPDATE STATISTICS FOR `travel-sample`._default.airport INDEX ((SELECT RAW name FROM system:indexes WHERE bucket_id = "travel-sample" AND scope_id = "_default" AND keyspace_id = "airport" AND state = "online" AND `using` = "gsi"));

Note that similar to BUILD INDEX statement, double parentheses
is required.

Change-Id: I9efcf7d2c5a5318c324abf87d38b711f393e86dd
Reviewed-on: http://review.couchbase.org/c/query/+/139868
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>