MB-42754 Evaluate additional ON-clause filters when joining on document key
When joining on document key, if other join methods are not feasible,
the server can convert the join into legacy lookup join. However, lookup
join does not currently have ability to evaluate any additional ON-clause
filters. For example, if we have:

route JOIN airline ON route.airline_id = meta(airline).id AND airline.country = "United States"

when converted to lookup join, only route.airline_id = meta(airline).id
is evaluated, the other parts of the ON-clause is currently NOT evaluated
which leads to incorrect result.

Add infrastructure to evalute additional ON-clause filters in lookup
join. This is only necessary when lookup join is the result of converting
from an ANSI JOIN query. Note this new filter is also different than
the filters in nested-loop join or hash join operators. In those cases
the extra filter is post-join, i.e. evaluated after the join. In case
of lookup join, the extra filter is part of ON-clause, i.e. evaluated
as part of the join operation itself. This difference is important in
case OUTER join is being used.

Change-Id: Iea126e58a1214326705be68b5ca39c7ab8816254
Reviewed-on: http://review.couchbase.org/c/query/+/140287
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>