MB-33009 Delay processing of query parameters till actual use
When a query contains named parameters and/or positional parameters,
and the value of the query parameters is known, a previous bug fix
(MB-24871) replaces the query paramter expression with its actual
value early on in the planning process, such that the rest of the
planner deals with constants instead of query parameters. While
this approach helps with certain aspects of planning such as
index where clause matching, it also causes issues since now we
have two different versions of the same expression, and this
mismatch can result in problems in covering determination, etc.

We now delay the processing of query parameters until it's
necessary to know the actual value of the query parameters,
including:
  - matching of index WHERE clause
  - matching of FTS index
  - determination of predicate selectivity (CBO)
  - input to index advisor
While the last one (index advisor) is not strictly necessary, it
makes coding much simpler since we now use a new interface to
represent the "context" for query-ee/optutil package, where
named/positional parameters can be retrieved. We pass in
PrepareContext from the planner for that purpose. The index
advisor already has an execution context passed in for purpose
of UPDATE STATISTICS and thus it's a bit tricker to try to pass
in a different context. The query parameters should not affect
index recommendations, and if CBO-based advise is being used
we use actual values to facilitate selectivity calculation.

Also updated unit test infrastructure to allow named and
positional parameters in explain checking as well.

Change-Id: I08f73cf5c5c7c3b27b1ea4d145a42b246175003b
Reviewed-on: http://review.couchbase.org/c/query/+/140398
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/query/+/146582
Well-Formed: Build Bot <build@couchbase.com>