MB-42855 Properly handle functions in generating index advice
When query predicate contains a function, e.g. lower(city) = 'lyon';
the normalization process turns it into lower((`travel-sample`.city)) = 'lyon';
Note the double parentheses, the outer set is for the function, and the inner
set is for a field. When an index key is generated we need to remove the
keyspace alias from the expression, currently this is done via
stringer.SetReplace(alias, "self", true). While this achieves the desired
result (i.e. removing keyspace alias), it leaves the double parentheses thus
the expression becomes lower((city)). However when a histogram is generated
it first call expression.TrimKeyspace() to remove the keyspace alias, thus
the histogram key is actually lower(city), (note the single set of
parentheses). This caused a mismatch in expression string and thus not able
to find the proper histogram, resulting in not able to use CBO-based advice.

The fix is to use the same call to function expression.TrimKeyspace() to
process the expression to ensure expression string matches the histogram
key.

Also added special handling of Meta() functions when trimming keyspace
alias.

Change-Id: Ib281ad52872f2ab6fd681644431bab62309279d4
Reviewed-on: http://review.couchbase.org/c/query/+/140666
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>