MB-42753 slow system:indexes and keyspaces with lots of collections
This defect is going to need loads of changes, including some architectural.
To get us started, introduce a 100ms threshold for namespace refreshes:
we don't want to bombard ns_server with GET /pools/default requests, when one suffices.
Also - add a preload flag to the Objects() method.
If we know we are going to use the object later, we might as well load the bucket now, than call ns_server twice.

Change-Id: I66ca756fa043c9ccf3af999eba8faa7ce8259770
Reviewed-on: http://review.couchbase.org/c/query/+/140709
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>