MB-42776 Same statement with only constant differences should get same index advice if run under Advisor() function
The rule-based index advisor may generate several index advices
for a statement, and randomly pick one of them for output. If
a session is started for Advisor() function and several such
statements are run, we may generate difference index advices
for similar statements that differ only in constants used.

To make it a bit more user-friendly, we detect situations when
an ADVISE statement is run under Advisor() function, and use
a seed for random number generator such that these similar
statements will get the same random number and thus the same
index advice.

A new flags field is added to execution context such that
we can mark the context when an Advisor() function is
encountered. Appropriate changes are also made for other
context interfaces that refers to the execution context.

The previous fix for MB-42445 is also partially reverted
since we now have a flag on the context for Advisor()
function.

Change-Id: I5d39c035ca7f29741b578cca8fdf8cebbbc83eff
Reviewed-on: http://review.couchbase.org/c/query/+/140724
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>