MB-42936 Report errors from Advise statement when run under Advisor() function
When an Advise statement is run under Advisor() function, either as
explicit argument or as part of a session (statements from
system:completed_requests), if an error is returned, currently the
error is not reported to the user, only advices from successful
Advise statements are returned.

Add infrastructure to keep track of errors from Advise statements
and report the errors as well to the user.

Change-Id: I66d7bf4ec59c90ceebd85774b15ebfe40ec850c7
Reviewed-on: http://review.couchbase.org/c/query/+/140929
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>