MB-42975 requests that panic don't report it
The issue here is as follows: all the defer this.notify() and this.close(context) that we have get called on panic and do a sendStop and close the operator
before context.Recover() gets a chance to process the panic.
This is dangerous because essentially we release the writer before the request has actually finished (wrong results, possibly more panics)

I have tried removing the defer this.notify() and close and move them at the end of all the runOnce() method,
but this entails changing 43 files and in may places changing the logic substantially, which is a good way to introduce bugs.

Golang doesn't offer a direct way to skip deferred functions, so this is a bit hacky, but does the job.

Change-Id: Ie64ab20795bc8adaa9be1f0f59379d228e8671fc
Reviewed-on: http://review.couchbase.org/c/query/+/141162
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>