MB-42956 Replace query parameters for generation of index spans for IN and LIKE clauses
The fix for MB-33009 delayed replacement of query parameters during
planning process, as a result, if a query paremeter is used in an
IN-list, we now generates a dynamic IN expansion instead of static
spans, and that in turn disables index aggregate pushdown since
the span is not exact.

To fix this, we now perform necessary query parameter replacement
when we generate index spans for IN-list, as well as LIKE clauses,
which are the two places currently where we could potentially set
exact to false. The other index span generations such as EQ, LT,
etc are not affected since we set exact to true regardless of
whether query parameters are used or not.

The unit test infrastructure is also updated to allow the option
of specifying whether query parameters should be used when
doing explain (which is needed for index WHERE clause matching
in case query parameter is used).

Change-Id: Ifc90eb0a31664539b24d7a5482552a998817c799
Reviewed-on: http://review.couchbase.org/c/query/+/141269
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>