MB-43088 Detect and recreate primary index for N1QL_CBO_STATS
If the primary index on N1QL_CBO_STATS is dropped, but the collection
is still in place, we assume this is a mistake and attempt to
recreate the index in the background such that next use of the index
will succeed. This should (hopefully) be a rare occurrence.

Also added error checking when system collection is being created
to detect rare cases where creation fails within the timeout period.

The creation of the primary index is also now independent of
the creation of the system collection, in the sense that if the
collection exists but the index does not, we'll create the index.

Change-Id: I981db513a4baf9dd0c5810881a3bff4fa22484aa
Reviewed-on: http://review.couchbase.org/c/query/+/141515
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>