MB-42577 Properly handle selectivities of LIKE predicates
LIKE predicates with leading wildcard cannot be costed properly.
It'll have a default selectivity. For index scan such LIKE
predicate can generate a span that includes all strings.
Currently there is no way to link such index span with the LIKE
predicate and thus it's not clear that the selectivity may have
been applied already.

Add new flags to mark the index span to indicate it's for LIKE
predicate with default selectivity. Then when marking index
filters also find and mark the LIKE predicate such that its
selectivity is not reapplied.

Also fixed an issue with early filtering where we apply the
transformed LIKE predicate but then remove the original LIKE
predicate from the FILTER operator later on. This is not always
correct. Instead we only remove the LIKE predicate if the
original LIKE predicate is being applied early.

Change-Id: I24f64eadab283a955eab2c8da611d277fb4033b8
Reviewed-on: http://review.couchbase.org/c/query/+/142435
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>