MB-32691 Detect nested correlation in FROM clause
Currently correlation is checked by references in the query block,
and marked on the SELECT. In the repro case we have nested subqueries
in the FROM clause. We correctly detected the inner most query block
as correlated, but not the intermediate subquery level since the
only correlation is in the FROM clause subquery which is currently
not detected.

Add a function to detect correlation in the FROM term interface
such that any correlated subquery or expression term in the FROM
clause can be easily determined, and set the correlation flag
in SELECT if such correlation is detected.

Change-Id: I2c9512ab7f7c79c58013af5d62162d6e1b0e192e
Reviewed-on: http://review.couchbase.org/c/query/+/142590
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>