MB-39397 Consider virtual indexes when index hints being used
When index hint is being used in an advise statement, for covering
index verification, currently only the indexes included in the
index hint is being considered, resulting in not able to verify
covering index (which is virtual index), and lack of covering
index advise.

The fix is to also include virtual indexes in the hint index
as well.

Change-Id: Idac4bcf9af4fe62d8cd3da93cd132c7cc12a8005
Reviewed-on: http://review.couchbase.org/c/query/+/142689
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>