MB-43436 Consolidate optimizer estimates in plan operators
Currently we have cost and cardinality as optimizer estimates
in plan operators. These are handled individually which makes
adding additional optimizer estimates cumbersome.

Add a new structure for optimizer estimates in plan and
embed this structure in plan operators, such that future
additions to optimizer estimates are easier to handle.

Change-Id: If288e78c706c1bd731c80226f0c33c85bb9b079a
Reviewed-on: http://review.couchbase.org/c/query/+/142696
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>