MB-43499 Add document size and first-result cost to plan operators
As a prerequisite for properly handle costing for LIMIT/OFFSET, we now
calculate a "first-result" cost for each operator, which represents
the cost for retrieving the first result from that operator.

In addition, a previous change requires document size(s) at each stage
for cost calculations. Current code recalculate the document sizes
based on the keyspaces that have been planned. Instead, we now keep
track of document sizes at each stage and build it up the plan operator
tree, so we don't need to recalculate at each operator. As a result of
this change, the previous fix for MB-43218 is no longer needed and thus
reversed.

Change-Id: Ib9153098d4e98150dfb0acd56294121be284bbbc
Reviewed-on: http://review.couchbase.org/c/query/+/142956
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>