MB-19881 Improve cost of logging
Removed *p and *m logging variants and added *a variants taking an anonymous function that returns a string as the argument.  Typical implementations will return fmt.Sprintf the advantage being that arguments aren't evaluated when not logging.  Tweaks to the level testing and formatters to improve performance.  Users of *p and *m variants altered to *f or *a counterparts.

Change-Id: I8a7246149c19390f92f8beb7eafc7251f984642c
Reviewed-on: http://review.couchbase.org/c/query/+/143539
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>