MB-37719 avoid returned strings escaping to heap
Initial changes for the Q26 hash join test slowdown.
With go 1.13 any function that returns a string as an interface{} copies it on the heap because the return value escapes (incorrectly, in my view).
This is of particular relevance for value.unwrap() and value.Actual().
Changed value comparison logic to use value.Type() and avoid most of the unwraps (it's leaner code anyway),
and added a ToString() conversion method as a quick Actual().(string) substitute.
(ToString() has a different use from String(), used by the stringer, and producing Marshalled strings).
There might be scope for a ToInt() and ToFloat() too.
Booleans are covered by Truth().
With this Q26 goes back to the same throughput, but slightly slower latency.

The remaining issue is that annotated and scope values used to be build the hash table are not recycled, and this still leans on the GC a lot.
I have tried recycling them as part of hashTable.Drop(), but this actually hinders throughput, so I'll have to find something better.

Change-Id: I47a52e4718c82e0bba8ce2ee19a665d4063a9894
Reviewed-on: http://review.couchbase.org/c/query/+/143708
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>