MB-37719 recycle group values take 2
Previous failure was due to a subquery recicling the parent value from the main query,
so introduced a way to stash and restore annotated values so that subqueries can't recycle it.
Also implemented more comprehensive group operator recycle.
I believe this should have the same throughput for Q26 as the previous change,
bit can't test it as Q26 is currently broken and doesn't run.

Change-Id: Icdb8eb4a96d2a447696bad502b8b8b5b15d1644f
Reviewed-on: http://review.couchbase.org/c/query/+/144259
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>