MB-42920 Allow meta().id to be included in advised index
Currently meta().id is excluded as an index key from recommended
indexes due to the fact that:
  - meta().id is already part of the index
  - we do not want to recommend an index that looks like a primary index

However in case of the customer's complaint an index with meta().id
as a non-leading index column could significantly improve query
performance if the query has a predicate on meta().id.

We now allow meta().id to be gathered if it appears in a predicate,
following the same rules as any other field in a predicate, such
that it can be included in recommended index. However, the following
cautions are taken:
  - do not include meta().id for covering purpose only
  - do not recommend an index with meta().id as the leading index
    key and there is no index condition (too similar to primary
    index in this case).

Change-Id: I59d2f886d873c4aed07e471cf1b42d33c3cb6ac6
Reviewed-on: http://review.couchbase.org/c/query/+/144399
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: http://review.couchbase.org/c/query/+/144562
Well-Formed: Build Bot <build@couchbase.com>