Update Index Advisor unit test
A previous change (when CBO-based index advisor was implemented) disabled
insertion of documents into the bucket before running ADVISE since at the
time CBO-based advise was triggered by non-0 number of documents in the
keyspace. This effectively disabled detection of flavor for partial index
since there is no infer possible with empty keyspace.

Add back the insertion of documents and updated index advise to now
include partial index with inferred flavor.

There is a pending issue with credentials in virtual index test and it's
temporarily disabled.

Change-Id: Ib96366afa9a1b0aff194c32bc572832560c26035
Reviewed-on: http://review.couchbase.org/c/query/+/144557
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>