MB-43739 Allow caching of subquery result for prepared statement
The main issue with this defect is currently for prepared statement
we are not caching (non-correlated) subquery result, thus the subquery
filter is executed repeatedly for each document fetched.

We now allow subquery result to be cached for prepared statement
as well. Since the cache is on the execution context and specific
to each execution of a statement, there should be no issue if the
statement is executed again (potentially with a different set of
query parameters).

A separate issue identified is that for ArrayConstruct and
ObjectConstruct expressions, since the PropagatesMissing() and
PropagatesNull() calls checks for this.value which is set when
Value() is called, the constructor for these expressions calls
Value() to intialize this.value. However this could later be reset
with ResetValue() call (e.g. when query parameters are being
replaced). We now explicitly initialize this.value by calling
Value() again as part of the reset.

Change-Id: Iec901117c9beb231c79e06934c188746dbb4e4a3
Reviewed-on: http://review.couchbase.org/c/query/+/146376
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>