MB-44247 truncate slices before pooling
QE can't test tonight and we've been asked to merge and they'll try tomorrow.
If this addresses the issue, the approach has to be extended to other slice pools.
The reason why the current code doesn't work is that the planner will remove duplicate entries from the index slice, leaving the pointers in place.
This goes through the whole buffer until it finds an empty slot.

Change-Id: Ia08a06b8dd877fed69ab1250b7f0ea0cba118126
Reviewed-on: http://review.couchbase.org/c/query/+/146459
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>